ai-startup-tracker/requirements.txt
anthropic>=0.40.0         # Claude SDK（主力 LLM）
litellm>=1.50.0           # LLM 代理层（方便换模型）
pydantic>=2.5.0           # 数据校验
pyairtable>=2.3.0         # Airtable SDK
beautifulsoup4>=4.12.0    # HTML 清洗
httpx>=0.27.0             # HTTP 客户端（Miniflux API）
tenacity>=8.2.0           # 重试
python-dotenv>=1.0.0      # 环境变量
